+++ src/Parser/Twig/TwigDocumentParser.php
@@ -6,12 +6,14 @@ use Symfony\Lsp\Parser\TreeSitter\TreeSitterParserInterface;
 
 final class TwigDocumentParser
 {
-    public function __construct(private readonly TreeSitterParserInterface $parser)
-    {
+    public function __construct(
+        private readonly TreeSitterParserInterface $parser,
+        private readonly TwigCommentParser $commentParser,
+    ) {
     }
 
     public function parse(string $source): TwigDocument
     {
-        return new TwigDocument($source, $this->parser->parse('twig', $source));
+        return new TwigDocument($source, $this->parser->parse('twig', $this->commentParser->mask($source)));
     }
 }