+++ tests/Feature/Route/TwigRouteReferenceExtractorTest.php
@@ -8,14 +8,20 @@ use Symfony\Lsp\Feature\Route\RouteReference;
 use Symfony\Lsp\Feature\Route\TwigRouteReferenceExtractor;
 use Symfony\Lsp\Parser\TreeSitter\NativeTreeSitterParser;
 use Symfony\Lsp\Parser\TreeSitter\TreeSitterResultDecoder;
+use Symfony\Lsp\Parser\Twig\TwigCommentParser;
 use Symfony\Lsp\Parser\Twig\TwigDocumentParser;
 
 final class TwigRouteReferenceExtractorTest extends TestCase
 {
     public function testExtractsPathAndUrlReferencesWithLiteralParameters(): void
     {
-        $references = (new TwigRouteReferenceExtractor(new PositionConverter(), new TwigDocumentParser(new NativeTreeSitterParser(new TreeSitterResultDecoder()))))->extract(<<<'TWIG'
+        $references = (new TwigRouteReferenceExtractor(new PositionConverter(), new TwigDocumentParser(new NativeTreeSitterParser(new TreeSitterResultDecoder()), new TwigCommentParser())))->extract(<<<'TWIG'
             {# {{ path('ignored') }} #}
+            {## {{ path('documented_outer') }} ##}
+            {% types {
+                ## path('documented')
+                article: 'string',
+            } %}
             {{ path(route_name('ignored')) }}
             <a href="{{ path('article_show', {'id': article.id}) }}">Show</a>
             <a href="{{ url("homepage") }}">Home</a>